Add history feature
You can now time-travel back to previous moves using buttons on the side.

The state is no longer one object that gets overwritten, but an array of states that gets concatenated or sliced to add new state or remove previous states. When timetraveling, the 'stepNumber' is the only thing initially changed, later states are only overwritten once a new move is made.