Add moves history sorting
This change introduces a new viewmodel version of the history state in Board's render. This new model adds a 'move' value which is the move index. This is redundant in the actual state object because the move index is implied by its position in the history array, however in this viewmodel we may want to re-order the history in which case order is no longer enough.